application updates

+++ app/Http/Controllers/API/UserController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\User;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Gate;
 
 class UserController extends Controller
 {
@@ -26,7 +27,10 @@ class UserController extends Controller
          */
     public function index()
     {
-        return User::latest()->paginate(10);
+        // $this->authorize('isAdmin');
+        if (Gate::allows('isAdmin') || Gate::allows('isModerator')) {
+            return User::latest()->paginate(4);
+        }
     }
 
     /**
@@ -57,7 +61,38 @@ class UserController extends Controller
     {
         $user = auth('api')->user();
 
-        return $request->photo;
+        $this->validate($request, [
+            'name' => 'required|string|max:191',
+            'email' => 'required|string|email|max:191|unique:users,email,' . $user->id,
+            'password' => 'sometimes|required|min:6'
+        ]);
+
+        $currentPhoto = $user->photo;
+
+        if ($request->photo != $currentPhoto) {
+            $name = time() . '.' . explode('/', explode(':', substr(
+                $request->photo,
+                0,
+            strpos($request->photo, ';')
+            ))[1])[1];
+
+            \Image::make($request->photo)->save(public_path('img/profile/') . $name);
+
+            $request->merge(['photo' => $name]);
+
+            $userPhoto = public_path('img/profile/') . $currentPhoto;
+
+            if (file_exists($userPhoto)) {
+                @unlink($userPhoto);
+            }
+        }
+
+        if (!empty($request->password)) {
+            $request->merge(['password' => Hash::make($request['password'])]);
+        }
+
+        $user->update($request->all());
+        return ['message' => 'Success'];
     }
 
     public function profile()
@@ -105,8 +140,24 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('isAdmin');
         $user = User::findOrFail($id);
         $user->delete();
         return['message' => 'User Deleted'];
+    }
+
+    public function search()
+    {
+        if ($search = \Request::get('q')) {
+            $users = User::where(function($query) use ($search) {
+                $query->where('name','LIKE',"%$search%")
+                      ->orWhere('email','LIKE',"%$search%")
+                      ->orWhere('level','LIKE',"%$search%");
+            })->paginate(10);
+        } else {
+            $users = User::latest()->paginate(4);
+        }
+
+        return $users;
     }
 }
